Add Euclidean Profile
EuclideanProfile is the first of a restructured profile type that is
designed to be more composable than the previous profiles. Rather than
taking in text, EuclideanProfile takes in numpy arrays representing the
extracted features from the text. Because of this change, profiles and
feature extractors can be used in different ways, which is necessary in
order to allow preprocessing of text before benchmarking.

To allow a flexible interface for Profile classes, the PaddedArray class
has been created. This class allows you to more conveniently deal with
collections of numpy arrays that have different lengths. Concrete
implementations of the Profile class are expected to work with
PaddedArray as input, whereas the abstract Profile class will convert
other types of arrays into an equivalent PaddedArray.

Also, mocks have been created for StyleProfile and TextProcessor, which
are needed for development on the backend.